Added a dialog for full name
Created a dialogfragment for use when users are logging
in for the first time. It is currently being called from the
main class and does nothing, but will soon be moved to
the loginActivity. Just pushing to mark some progress.

Names entered are logged with the tag "Accounts" if
the "Let's Go!" button is pressed.