Authentication working
Authentication is now working, but unfortunately as we
currently have it, it is impossible for a user to login if they
change their UCSD AD password. This is due mostly to the
fact that passwords cannot be changed in Firebase
without first logging the user in.

Custom OAuth tokens are possible, but only if I was
running my own server. Therefore we will probably need
to scrap the LDAP authentication and roll out a purely
Firebase auth system. It adds one additional step in that
users need to create a password.