Added sliding animations
Sliding animations added to the login/signup screens so
that the screen will slide left when the create account
page is opened and slide right when it is closed.

Also updated dependencies to be much lighter here to
reduce the size of the application and add some
backwards compatibility to older api's.

Updated vector graphic (banana on front page) to use
support library. This won't matter in the near future as
we will probably not be using vectors anyways, but in
any case it adds more backwards-compatibility with the
likes of Marshmallow.