Added fragments for group joining and creation
Added fragments for group joining and creation, as well
as listeners to the buttons on the home page fragment.
Buttons will open both of these group joining and creation
pages. Animations have been implemented for both as
well.

Still need to switch menu button to a back button when
not on the home page, and add first and last name fields
for the account creation page.