Removed class loading dialog
Class loading dialog is largely useless, as the loading can
be done in the background. The array adapter is attached
on view creation and dynamically updated later, so I
**think** null pointers won't be an issue.

Slight bug: If the autocorrectview is focused before the
course data is fetched, it will never show completions
while it is still focused.