Adding and removing classes implemented
Users can now add and remove classes from the
expandable list view. Buttons subject to change of course.
Rather messy code at the moment, needs review. Overall
the interface seems to be smooth so that is a good sign.

User profiles are now created on account creation.
Note that all user accounts were deleted around the
time of this commit, to ensure profiles are created for
all users.