Finished join group children view
Need to add onclick listeners for the location and join
group buttons. Groups are updated live, and seem to be
functioning well. One known bug so far: removing a group
that is in the middle of the arraylist of groups will cause
a runtime exception when the list view of groups is
redrawn. I suspect that something is not getting updated
somewhere when a group is removed.