Group leader can now edit group
Group leader now has ability to edit the group with the action button
in the top right of the group interaction page. Changes are updated
to the database once the save button is pressed.

Still need to add custom array adapter to view and kick members,
as well as transfer leadership to them.