Added a background service
Background service will now run while the user is a member of a
group. It will persist even if the app is closed. It will later be used to
terminate the group at end time, send notifications, and update
details about the group when the app is not in the foreground.