Network connectivity issue fixed
Added connection state listener to Firebase. Firebase will now check
to ensure that it is connected before the user is allowed to use the
app. When the network goes down, the user is shown a spinning
progress indicator until it comes back up. Very rudimentary, but
effective for the most part.

Minor bug: when the user is disconnected (the progress
indicator is spinning) and the user reopens the activity by
pressing on the icon, the progress indicator will freeze onto the
screen, forcing the user to close the app and reopen via the icon.
Definitely needs to be fixed but is a low priority at the moment.