Fixed firebase disconnection bug
If left idle on the home screen, the app would show the disconnection
message after one minute. This seems to be due to a firebase
implementation. The disconnect does not happen when there are
active firebase listeners, so I simply added a listener to the user's
root in the database. Will continue to test to ensure this does not
introduce more bugs.