Groups now automatically destroyed
Created a broadcast receiver and register the end time with the
alarmManager. Receiver gets called and deletes the group when the
end time is reached. Still need to add a popup 5 minutes before the
expiry. Phone shutdown should also eventually remove the user
from the group and/or destroy the group.