Made secret button able to play music throughout app
Secret button is READY TO ROLL. Just needs a new, custom song. Will stop playing song when app is killed, or secret button is in the off state.
To do some amount of years in future: Create a way to pass through the media player to activities, and stop sound once the app is exited.